Change/fix source_id which is 32-bit in zmq messages v2

+++ include/ZMQCollectorInterface.h
@@ -36,7 +36,7 @@ typedef struct {
 class ZMQCollectorInterface : public ZMQParserInterface {
  private:
   void *context;
-  std::map<u_int8_t, u_int32_t> source_id_last_msg_id;
+  std::map<u_int32_t, u_int32_t> source_id_last_msg_id;
   bool is_collector;
   u_int16_t num_subscribers;
   zmq_subscriber subscriber[MAX_ZMQ_SUBSCRIBERS];
